+++ src/objects/shading.c
@@ -6,7 +6,7 @@
 /*   By: iherman- <iherman-@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/19 11:00:34 by johnhapke         #+#    #+#             */
-/*   Updated: 2025/10/28 11:43:46 by iherman-         ###   ########.fr       */
+/*   Updated: 2025/10/28 17:44:08 by iherman-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,55 +39,50 @@ static bool	reaches_light(const t_hitinfo *prev_hitinfo, t_rt_data *data)
 	return (true);
 }
 
-t_vec3	calculate_specular(const t_hitinfo *hitinfo, const t_rt_data *data)
-{
-	t_vec3	l;
-	double	ndotl;
-	double	rdotv;
-	double	specular_intensity;
-
-	if (SPECULAR_M == -1)
-		return ((t_vec3){0,0,0});
-	l = normalize(vector_subtract(data->light.light_point,
-				hitinfo->pos));
-	ndotl = fmax(0, vector_dot(hitinfo->surface_dir, l));
-	rdotv = fmax(0, vector_dot(vector_subtract
-				(vector_multiply(hitinfo->surface_dir, 2 * ndotl), l),
-				normalize(vector_subtract(data->camera.viewpoint,
-						hitinfo->pos))));
-	specular_intensity = pow(rdotv, SPECULAR_M);
-	return (vector_multiply((t_vec3){255, 255, 255}, specular_intensity));
-}
-
 unsigned int	format_color(t_vec3 color)
 {
-	return ((int)color.x << 24 | (int)color.y << 16 | (int)color.z << 8 | 255);
+	return ((int)(fmin(1.0, color.x) * 255.0) << 24
+		| (int)(fmin(1.0, color.y) * 255.0) << 16
+		| (int)(fmin(1.0, color.z) * 255.0) << 8
+		| 255);
+}
+
+t_vec3	calculate_ambient(const t_hitinfo *hitinfo, t_rt_data *data)
+{
+	t_vec3	ambient_color;
+
+	ambient_color.x = fmin(1.0, ((hitinfo->obj_color.x * data->amb_light.color.x) * data->amb_light.light_ratio));
+	ambient_color.y = fmin(1.0, ((hitinfo->obj_color.y * data->amb_light.color.y) * data->amb_light.light_ratio));
+	ambient_color.z = fmin(1.0, ((hitinfo->obj_color.z * data->amb_light.color.z) * data->amb_light.light_ratio));
+	return (ambient_color);
+}
+
+static inline t_vec3 vector_hadamard(t_vec3 a, t_vec3 b)
+{
+    t_vec3 res;
+
+    res.x = a.x * b.x;
+    res.y = a.y * b.y;
+    res.z = a.z * b.z;
+    return res;
 }
 
 unsigned int	calculate_color(const t_hitinfo *hitinfo, t_rt_data *data)
 {
-	t_vec3	ambient_color;
-	t_vec3	diffuse_color;
-	t_vec3	specular_color;
-	t_vec3	final_color;
-	double	diffuse_intensity;
+	const t_vec3	ambient_color = calculate_ambient(hitinfo, data);
+	t_vec3			diffuse_color;
+	t_vec3			final_color;
+	t_vec3			light_dir;
+	double			diffuse_intensity;
 
-	ambient_color = vector_multiply(hitinfo->obj_color,
-			data->amb_light.light_ratio);
+	final_color = ambient_color;
 	if (reaches_light(hitinfo, data))
 	{
-		diffuse_intensity = fmax(0, vector_dot(hitinfo->surface_dir,
-					normalize(vector_subtract(data->light.light_point,
-							hitinfo->pos))) * data->light.light_ratio);
-		diffuse_color = vector_multiply(hitinfo->obj_color, diffuse_intensity);
-		specular_color = calculate_specular(hitinfo, data);
-		final_color.x = (int)fmin(255.0, ambient_color.x
-				+ diffuse_color.x + specular_color.x);
-		final_color.y = (int)fmin(255.0, ambient_color.y
-				+ diffuse_color.y + specular_color.y);
-		final_color.z = (int)fmin(255.0, ambient_color.z
-				+ diffuse_color.z + specular_color.z);
-		return (format_color(final_color));
+		light_dir = normalize(vector_subtract(data->light.light_point, hitinfo->pos));
+		diffuse_intensity = fmax(0.0, vector_dot(hitinfo->surface_dir, light_dir)) * data->light.light_ratio;
+		diffuse_color = vector_hadamard(hitinfo->obj_color, data->light.color);
+		diffuse_color = vector_multiply(diffuse_color, diffuse_intensity);
+		final_color = vector_add(final_color, diffuse_color);
 	}
-	return (format_color(ambient_color));
+	return (format_color(final_color));
 }